+++ img2xl_app/models/notification.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from djangae import models
+from django.db import models
 from django.contrib.auth.models import User
 
 import logging
