+++ img2xl_app/models/usage_log.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from djangae import models
+from django.db import models
 from django.contrib.auth.models import User
 from django.utils import timezone
