+++ img2xl_app/models/user_profile.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from djangae import models
+from django.db import models
 from django.contrib.auth.models import User
 from django.db.models.signals import post_save
 from django.dispatch import receiver
